Move activity type to shared types + Restructure package
