+++ packages/types/src/generic.ts
@@ -1,0 +1,1 @@
+export type Units = 'metric' | 'imperial';
